pyproject.toml
[project]
name = "bitewise"
version = "0.1.0"
description = ""
authors = [
    {name = "Max Li",email = "143444069+notttired@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "recipe-scrapers (>=15.7.1,<16.0.0)",
    "pymongo (>=4.12.1,<5.0.0)",
    "openai-agents (>=0.0.14,<0.0.15)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "asynchttp (>=0.0.4,<0.0.5)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
